🚀 feat(.gitignore): add .vscode/ to gitignore 🎉 feat(bte.py): add support for BLE message receiving 🎉 feat(hdc1080.py): add driver for HDC1080 temperature and humidity sensor 🎉 feat(main.py): add support for HDC1080, VEML6040, Moisture and Battery sensors 🎉 feat(moisture.py): add driver for moisture sensor 🎉 feat(mqtt.py): add MQTT client class Added .vscode/ to gitignore to avoid committing vscode settings. Added support for receiving BLE messages in bte.py. Added driver for HDC1080 temperature and humidity sensor in hdc1080.py. Added support for HDC1080, VEML6040, Moisture and Battery sensors in main.py. Added driver for moisture sensor in moisture.py. Added MQTT client class in mqtt.py.
🎉 feat(veml6040.py): add VEML6040 class to read data from the VEML6040 i2c light sensor
The VEML6040 class reads RGB or HSV data and classifies Hue given a dictionary of colour definitions. The class also includes a method to convert RGB to HSV.

🎉 feat(wifi.py): add Wifi class to connect to wifi network
The Wifi class reads wifi connection details from a file and connects to the wifi network. The class also includes a method to check if the device is connected to the wifi network.